convert uint64 to int64 for BSON no support uint64

+++ modules/oracle/edit_feed.go
@@ -7,13 +7,13 @@ import (
 
 type DocMsgEditFeed struct {
 	FeedName          string        `bson:"feed_name" yaml:"feed_name"`
-	LatestHistory     uint64        `bson:"latest_history" yaml:"latest_history"`
+	LatestHistory     int64         `bson:"latest_history" yaml:"latest_history"`
 	Description       string        `bson:"description"`
 	Creator           string        `bson:"creator"`
 	Providers         []string      `bson:"providers"`
 	Timeout           int64         `bson:"timeout"`
 	ServiceFeeCap     []models.Coin `bson:"service_fee_cap" yaml:"service_fee_cap"`
-	RepeatedFrequency uint64        `bson:"repeated_frequency" yaml:"repeated_frequency"`
+	RepeatedFrequency int64         `bson:"repeated_frequency" yaml:"repeated_frequency"`
 	ResponseThreshold uint32        `bson:"response_threshold" yaml:"response_threshold"`
 }
 
@@ -25,13 +25,13 @@ func (m *DocMsgEditFeed) BuildMsg(v interface{}) {
 	msg := v.(*MsgEditFeed)
 
 	m.FeedName = msg.FeedName
-	m.LatestHistory = msg.LatestHistory
+	m.LatestHistory = int64(msg.LatestHistory)
 	m.Description = msg.Description
 	m.Creator = msg.Creator
 	m.Providers = msg.GetProviders()
 	m.Timeout = msg.Timeout
 	m.ServiceFeeCap = models.BuildDocCoins(msg.ServiceFeeCap)
-	m.RepeatedFrequency = msg.RepeatedFrequency
+	m.RepeatedFrequency = int64(msg.RepeatedFrequency)
 	m.ResponseThreshold = msg.ResponseThreshold
 }
